requirements.txt
pygame==2.0.1
tensorflow==2.9.1
pillow==8.4.0
numpy==1.21.2